計算.pyで緩和係数を下げないオプションが動作するようにした

+++ binDEXCS（解析フォルダを端末で開いてから）/improveMeshQualityを実行.py
@@ -19,8 +19,8 @@ from struct import pack
 from utilities import misc
 from utilities import rmObjects
 
-if __name__ == '__main__':
-    signal.signal(signal.SIGINT, signal.SIG_DFL) # Ctrl+Cで終了
+if __name__ == "__main__":
+    signal.signal(signal.SIGINT, signal.SIG_DFL)  # Ctrl+Cで終了
     misc.showDirForPresentAnalysis(__file__)
 
     if len(sys.argv) == 1:
@@ -31,74 +31,91 @@ if __name__ == '__main__':
         scaleMesh_0p001 = False
         i = 1
         while i < len(sys.argv):
-            if sys.argv[i] == '-N': # Non-interactive
+            if sys.argv[i] == "-N":  # Non-interactive
                 pass
-            elif sys.argv[i] == '-a':
+            elif sys.argv[i] == "-a":
                 i += 1
                 patch_name = sys.argv[i]
-            elif sys.argv[i] == '-p':
+            elif sys.argv[i] == "-p":
                 exec_paraFoam = True
-            elif sys.argv[i] == '-s':
+            elif sys.argv[i] == "-s":
                 scaleMesh_0p001 = True
             i += 1
 
-    faces_path = os.path.join('constant', 'polyMesh', 'faces')
-    for i in (os.path.join('constant', 'polyMesh', 'boundary'), faces_path):
+    faces_path = os.path.join("constant", "polyMesh", "faces")
+    for i in (os.path.join("constant", "polyMesh", "boundary"), faces_path):
         if not os.path.isfile(i):
-            print(f'エラー: {i}ファイルがありません．')
+            print(f"エラー: {i}ファイルがありません．")
             sys.exit(1)
-    if os.path.isdir('dynamicCode'):
-        shutil.rmtree('dynamicCode')
+    if os.path.isdir("dynamicCode"):
+        shutil.rmtree("dynamicCode")
     converted_millimeter_into_meter = misc.isConvertedMillimeterIntoMeter()
 
-    with open(faces_path, 'r') as f:
+    with open(faces_path, "r") as f:
         s = f.read()
-    if 'faceCompactList' in s:
-        pl = s.find('(') + 1
-        ph = s[:pl - 2].rfind('\n') + 1
-        pf = s.rfind(')')
-        n = int(s[ph:pl - 1])
-        pr = pl + 4*n
-        faces_ranges = np.fromstring(s[pl:pr], dtype = '<u4')
-        pl = pr + s[pr:].find(')') + 1
-        pr = pl + s[pl:].find('(')
+    if "faceCompactList" in s:
+        pl = s.find("(") + 1
+        ph = s[: pl - 2].rfind("\n") + 1
+        pf = s.rfind(")")
+        n = int(s[ph : pl - 1])
+        pr = pl + 4 * n
+        faces_ranges = np.fromstring(s[pl:pr], dtype="<u4")
+        pl = pr + s[pr:].find(")") + 1
+        pr = pl + s[pl:].find("(")
         n = int(s[pl:pr])
         pl = pr + 1
-        pr = pl + 4*n
-        points_indices = np.fromstring(s[pl:pr], dtype = '<u4')
+        pr = pl + 4 * n
+        points_indices = np.fromstring(s[pl:pr], dtype="<u4")
         faces_data = []
         for i in range(faces_ranges.shape[0] - 1):
-            faces_data.append(points_indices[faces_ranges[i]:faces_ranges[i + 1]])
-        os.rename(faces_path, f'{faces_path}_bak')
-        with open(faces_path, 'w') as f:
-            f.write(s[:ph].replace('faceCompactList', 'faceList'))
-            f.write(f'{len(faces_data)}\n(\n')
+            faces_data.append(points_indices[faces_ranges[i] : faces_ranges[i + 1]])
+        os.rename(faces_path, f"{faces_path}_bak")
+        with open(faces_path, "w") as f:
+            f.write(s[:ph].replace("faceCompactList", "faceList"))
+            f.write(f"{len(faces_data)}\n(\n")
             for i in faces_data:
-                f.write(f'{i.shape[0]}(')
+                f.write(f"{i.shape[0]}(")
                 for j in i:
-                    f.write(pack('<d', j))
-                f.write(')\n')
+                    f.write(pack("<d", j))
+                f.write(")\n")
             f.write(s[pf:])
 
-    if misc.execCommand(['improveMeshQuality', '-noFunctionObjects'])[1] != 0:
+    if misc.execCommand(["improveMeshQuality", "-noFunctionObjects"])[1] != 0:
         sys.exit(1)
 
     if converted_millimeter_into_meter:
         misc.convertMillimeterIntoMeter()
     else:
         if interactive:
-            box = misc.bounding_box_of_calculation_range(os.path.join('constant', 'polyMesh', 'points'))[1]
-            print(f'元のメッシュの範囲は{box[0][0]} <= x <= {box[0][1]}, {box[1][0]} <= y <= {box[1][1]}, '
-                f'{box[2][0]} <= z <= {box[2][1]}です．')
-            scaleMesh_0p001 = True if input('この長さの単位はミリメートルですか？ '
-                '(y/n, yだと0.001倍してメートルに直します．) > ').strip().lower() == 'y' else False
+            box = misc.bounding_box_of_calculation_range(
+                os.path.join("constant", "polyMesh", "points")
+            )[1]
+            print(
+                f"元のメッシュの範囲は{box[0][0]} <= x <= {box[0][1]}, {box[1][0]} <= y <= {box[1][1]}, "
+                f"{box[2][0]} <= z <= {box[2][1]}です．"
+            )
+            scaleMesh_0p001 = (
+                True
+                if input(
+                    "この長さの単位はミリメートルですか？ "
+                    "(y/n, yだと0.001倍してメートルに直します．) > "
+                )
+                .strip()
+                .lower()
+                == "y"
+                else False
+            )
         if scaleMesh_0p001:
             misc.convertMillimeterIntoMeter()
-    misc.removePatchesHavingNoFaces() # フェイスを1つも含まないパッチを取り除く
+    misc.removePatchesHavingNoFaces()  # フェイスを1つも含まないパッチを取り除く
     misc.execCheckMesh()
 
     if interactive:
-        exec_paraFoam = True if input('\nparaFoamを実行しますか？ (y/n) > ').strip().lower() == 'y' else False
-    misc.execParaFoam(touch_only = not exec_paraFoam, ambient = 0.0, diffuse = 1.0)
+        exec_paraFoam = (
+            True
+            if input("\nparaFoamを実行しますか？ (y/n) > ").strip().lower() == "y"
+            else False
+        )
+    misc.execParaFoam(touch_only=not exec_paraFoam, ambient=0.0, diffuse=1.0)
 
     rmObjects.removeInessentials()
